removing extra RstudioMainPage
