+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/RuntimeFixtureSpec.scala
@@ -21,7 +21,8 @@ abstract class RuntimeFixtureSpec
     extends FixtureAnyFreeSpec
     with BeforeAndAfterAll
     with LeonardoTestUtils
-    with Retries {
+    with Retries
+    with GPAllocBeforeAndAfterAll {
 
   implicit val ronToken: AuthToken = ronAuthToken
