+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/rstudio/RStudioSpec.scala
@@ -3,7 +3,7 @@ package org.broadinstitute.dsde.workbench.leonardo.rstudio
 import org.broadinstitute.dsde.workbench.leonardo._
 import org.openqa.selenium.{Keys}
 import org.scalatest.DoNotDiscover
-@DoNotDiscover
+//@DoNotDiscover
 class RStudioSpec extends RuntimeFixtureSpec with RStudioTestUtils {
 
   override val toolDockerImage: Option[String] = Some(LeonardoConfig.Leonardo.rstudioBioconductorImage.imageUrl)
